Task 4 typescript basics

+++ src/resources/boards/board.router.ts
@@ -1,13 +1,14 @@
-const router = require('express').Router();
-const boardService = require('./board.service');
-const Board = require('./board.model');
-const { TASKS } = require('../tasks/task.memory.repository');
-const tasksService = require('../tasks/task.service');
+import express from 'express';
+import * as  boardService from './board.service';
+import { Board } from './board.model';
+import * as tasksService from '../tasks/task.service';
+
+const router = express.Router();
 
 /**
  * GET All boards
  */
-router.route('/').get(async (req, res) => {
+router.route('/').get(async (_, res) => {
   const boards = await boardService.getAll();
   res.json(boards.map(Board.toResponse));
 });
@@ -29,7 +30,12 @@ router.route('/').post(async (req, res) => {
  * GET board by id
  */
 router.route('/:id').get(async (req, res) => {
-  const board = await boardService.getBoardById(req.params.id);
+  let board;
+  const { id: boardId } = req.params;
+
+  if (boardId) {
+    board = await boardService.getBoardById(boardId);
+  }
 
   if (board) {
     res.status(200).send(Board.toResponse(board));
@@ -42,7 +48,12 @@ router.route('/:id').get(async (req, res) => {
  * Update (PUT) board
  */
 router.route('/:id').put(async (req, res) => {
-  const board = await boardService.updateBoard(req.params.id, req.body);
+  let board;
+  const { id: boardId } = req.params;
+
+  if (boardId) {
+    board = await boardService.updateBoard(boardId, req.body);
+  }
 
   if (board) {
     res.status(200).send(Board.toResponse(board));
@@ -56,20 +67,20 @@ router.route('/:id').put(async (req, res) => {
  */
 router.route('/:id').delete(async (req, res) => {
   try {
-    const boardId = req.params.id;
+    const { id: boardId } = req.params;
 
-    // deleting board
-    await boardService.deleteBoardById(boardId);
+    if (boardId) {
+      // deleting board
+      await boardService.deleteBoardById(boardId);
 
-    // deleting tasks
-    const tasksForDeleting = TASKS.filter(task => task.boardId === boardId);
+      // deleting tasks for particular board
+      await tasksService.deleteTasksForParticularBoardId(boardId);
 
-    await Promise.all(tasksForDeleting.map((task) => tasksService.deleteTaskById(boardId, task.id)));
-
-    res.status(204).send('Board has been deleted');
+      res.status(204).send('Board has been deleted');
+    }
   } catch (e) {
     res.status(404).end('Board has not been deleted');
   }
 });
 
-module.exports = router;
+export default router;
